Fixed format specifier warnings for 32/64 bit build

+++ NSScanner+Fragaria.m
@@ -25,7 +25,7 @@
      */
     NSUInteger maxIndex = [[self string] length];
 	if (idx > maxIndex) {
-        NSLog(@"Invalid scan location %u > max of %u", idx, maxIndex);
+        NSLog(@"Invalid scan location %lu > max of %lu", (long)idx, (long)maxIndex);
 		idx = maxIndex;
 	}
